requirements.txt
flask
torch
torchvision
Pillow
numpy
opencv-python
kaggle
tqdm
tensorflow
requests
scikit-learn
